:truck: Move global error types to shared
