+++ src/Changelogs/AuthManager.ts
@@ -1,6 +1,6 @@
 import { UserToken } from '../shared/types';
 import { DataStore } from '../shared/interfaces/DataStore';
-import { ResourceError, ResourceErrorReason } from '../errors';
+import { ResourceError, ResourceErrorReason } from '../shared/errors';
 
 
 const ROLE = {
