+++ src/LearningObjectSubmission/SubmissionRouteDriver.ts
@@ -2,7 +2,7 @@ import { Request, Response, Router } from 'express';
 import { DataStore } from '../shared/interfaces/DataStore';
 import { FileManager } from '../shared/interfaces/interfaces';
 import { cancelSubmission, submitForReview, checkFirstSubmission } from './SubmissionInteractor';
-import { mapErrorToResponseData } from '../errors';
+import { mapErrorToResponseData } from '../shared/errors';
 import { UserToken } from '../shared/types';
 
 /**
