+++ src/LearningObjects/AuthorizationManager.ts
@@ -2,7 +2,7 @@
  * Set of reusable functions used to authorize requests within this module
  */
 
-import { ResourceError, ResourceErrorReason } from '../errors';
+import { ResourceError, ResourceErrorReason } from '../shared/errors';
 import { UserToken } from '../shared/types';
 
 enum AccessGroup {
